Simplify structure and remove unused code
